+++ ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/fetching.h
@@ -42,21 +42,6 @@ public:
 
 class IDataSource;
 
-class TApplyIndexStep: public IFetchingStep {
-private:
-    using TBase = IFetchingStep;
-    const NIndexes::TIndexCheckerContainer IndexChecker;
-
-protected:
-    virtual TConclusion<bool> DoExecuteInplace(const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& step) const override;
-
-public:
-    TApplyIndexStep(const NIndexes::TIndexCheckerContainer& indexChecker)
-        : TBase("APPLY_INDEX")
-        , IndexChecker(indexChecker) {
-    }
-};
-
 class TDetectInMemStep: public IFetchingStep {
 private:
     using TBase = IFetchingStep;
@@ -152,26 +137,6 @@ protected:
 public:
     TPortionAccessorFetchingStep()
         : TBase("FETCHING_ACCESSOR") {
-    }
-};
-
-class TIndexBlobsFetchingStep: public IFetchingStep {
-private:
-    using TBase = IFetchingStep;
-    std::shared_ptr<TIndexesSet> Indexes;
-
-protected:
-    virtual TConclusion<bool> DoExecuteInplace(const std::shared_ptr<IDataSource>& source, const TFetchingScriptCursor& step) const override;
-    virtual TString DoDebugString() const override {
-        return TStringBuilder() << "indexes=" << Indexes->DebugString() << ";";
-    }
-
-public:
-    TIndexBlobsFetchingStep(const std::shared_ptr<TIndexesSet>& indexes)
-        : TBase("FETCHING_INDEXES")
-        , Indexes(indexes) {
-        AFL_VERIFY(Indexes);
-        AFL_VERIFY(Indexes->GetIndexesCount());
     }
 };
