+++ ydb/core/tx/columnshard/engines/reader/simple_reader/iterator/source.cpp
@@ -224,86 +224,7 @@ TConclusion<NArrow::TColumnFilter> TPortionDataSource::DoCheckIndex(
     return filter.And(GetStageData().GetTable()->GetFilter());
 }
 
-bool TPortionDataSource::DoStartFetchingIndexes(
-    const std::shared_ptr<IDataSource>& sourcePtr, const TFetchingScriptCursor& step, const std::shared_ptr<TIndexesSet>& indexes) {
-    AFL_DEBUG(NKikimrServices::TX_COLUMNSHARD_SCAN)("event", step.GetName());
-    AFL_VERIFY(indexes->GetIndexesCount());
-    AFL_DEBUG(NKikimrServices::TX_COLUMNSHARD_SCAN)("event", step.GetName())("fetching_info", step.DebugString());
-
-    TBlobsAction action(GetContext()->GetCommonContext()->GetStoragesManager(), NBlobOperations::EConsumer::SCAN);
-    {
-        std::set<ui32> indexIds;
-        for (auto&& i : GetStageData().GetPortionAccessor().GetIndexesVerified()) {
-            if (!indexes->GetIndexIdsSet().contains(i.GetIndexId())) {
-                continue;
-            }
-            indexIds.emplace(i.GetIndexId());
-            if (auto bRange = i.GetBlobRangeOptional()) {
-                auto readAction = action.GetReading(Portion->GetIndexStorageId(i.GetIndexId(), Schema->GetIndexInfo()));
-                readAction->SetIsBackgroundProcess(false);
-                readAction->AddRange(Portion->RestoreBlobRange(*bRange));
-            }
-        }
-        if (indexes->GetIndexIdsSet().size() != indexIds.size()) {
-            return false;
-        }
-    }
-    auto readingActions = action.GetReadingActions();
-    if (!readingActions.size()) {
-        NYDBTest::TControllers::GetColumnShardController()->OnIndexSelectProcessed({});
-        return false;
-    }
-
-    auto constructor =
-        std::make_shared<NCommon::TBlobsFetcherTask>(readingActions, sourcePtr, step, GetContext(), "CS::READ::" + step.GetName(), "");
-    NActors::TActivationContext::AsActorContext().Register(new NOlap::NBlobOperations::NRead::TActor(constructor));
-    return true;
-}
-
 void TPortionDataSource::DoAbort() {
-}
-
-void TPortionDataSource::DoApplyIndex(const NIndexes::TIndexCheckerContainer& indexChecker) {
-    THashMap<NIndexes::TIndexDataAddress, std::vector<TString>> indexBlobs;
-    const auto indexIds = NIndexes::TIndexDataAddress::ExtractIndexIds(indexChecker->GetIndexIds());
-    //    NActors::TLogContextGuard gLog = NActors::TLogContextBuilder::Build()("records_count", GetRecordsCount())("portion_id", Portion->GetPortionId());
-    std::vector<TPortionDataAccessor::TPage> pages = GetStageData().GetPortionAccessor().BuildPages();
-    NArrow::TColumnFilter constructor = NArrow::TColumnFilter::BuildAllowFilter();
-    for (auto&& p : pages) {
-        for (auto&& i : p.GetIndexes()) {
-            if (!indexIds.contains(i->GetIndexId())) {
-                continue;
-            }
-            if (i->HasBlobData()) {
-                indexBlobs[NIndexes::TIndexDataAddress(i->GetIndexId())].emplace_back(i->GetBlobDataVerified());
-            } else {
-                indexBlobs[NIndexes::TIndexDataAddress(i->GetIndexId())].emplace_back(StageData->ExtractBlob(i->GetAddress()));
-            }
-        }
-        for (auto&& i : indexChecker->GetIndexIds()) {
-            if (!indexBlobs.contains(i)) {
-                GetContext()->GetCommonContext()->GetCounters().OnNotIndexBlobs();
-                return;
-            }
-        }
-        if (indexChecker->Check(indexBlobs)) {
-            NYDBTest::TControllers::GetColumnShardController()->OnIndexSelectProcessed(true);
-            constructor.Add(true, p.GetRecordsCount());
-            GetContext()->GetCommonContext()->GetCounters().OnAcceptedByIndex(p.GetRecordsCount());
-        } else {
-            NYDBTest::TControllers::GetColumnShardController()->OnIndexSelectProcessed(false);
-            constructor.Add(false, p.GetRecordsCount());
-            GetContext()->GetCommonContext()->GetCounters().OnDeniedByIndex(p.GetRecordsCount());
-        }
-    }
-    AFL_VERIFY(constructor.GetRecordsCountVerified() == Portion->GetRecordsCount());
-    if (constructor.IsTotalDenyFilter()) {
-        StageData->AddFilter(NArrow::TColumnFilter::BuildDenyFilter());
-    } else if (constructor.IsTotalAllowFilter()) {
-        return;
-    } else {
-        StageData->AddFilter(constructor);
-    }
 }
 
 TConclusion<bool> TPortionDataSource::DoStartFetchData(
