+++ ydb/core/tx/columnshard/engines/scheme/indexes/abstract/checker.h
@@ -63,46 +63,4 @@ public:
     }
 };
 
-class IIndexChecker {
-protected:
-    virtual bool DoCheck(const THashMap<TIndexDataAddress, std::vector<TString>>& blobs) const = 0;
-    virtual bool DoDeserializeFromProto(const NKikimrSSA::TProgram::TOlapIndexChecker& proto) = 0;
-    virtual void DoSerializeToProto(NKikimrSSA::TProgram::TOlapIndexChecker& proto) const = 0;
-    virtual std::set<TIndexDataAddress> DoGetIndexIds() const = 0;
-
-public:
-    using TFactory = NObjectFactory::TObjectFactory<IIndexChecker, TString>;
-    using TProto = NKikimrSSA::TProgram::TOlapIndexChecker;
-    virtual ~IIndexChecker() = default;
-    bool Check(const THashMap<TIndexDataAddress, std::vector<TString>>& blobs) const {
-        return DoCheck(blobs);
-    }
-
-    bool DeserializeFromProto(const NKikimrSSA::TProgram::TOlapIndexChecker& proto) {
-        return DoDeserializeFromProto(proto);
-    }
-
-    void SerializeToProto(NKikimrSSA::TProgram::TOlapIndexChecker& proto) const {
-        return DoSerializeToProto(proto);
-    }
-
-    std::set<TIndexDataAddress> GetIndexIds() const {
-        return DoGetIndexIds();
-    }
-
-    virtual TString GetClassName() const = 0;
-};
-
-class TIndexCheckerContainer: public NBackgroundTasks::TInterfaceProtoContainer<IIndexChecker> {
-private:
-    using TBase = NBackgroundTasks::TInterfaceProtoContainer<IIndexChecker>;
-
-public:
-    TIndexCheckerContainer() = default;
-    TIndexCheckerContainer(const std::shared_ptr<IIndexChecker>& object)
-        : TBase(object) {
-        AFL_VERIFY(Object);
-    }
-};
-
 }   // namespace NKikimr::NOlap::NIndexes
