--- ydb/core/tx/columnshard/engines/scheme/indexes/abstract/composite.cpp
@@ -1,5 +1,0 @@
-#include "composite.h"
-
-namespace NKikimr::NOlap::NIndexes {
-
-}   // namespace NKikimr::NOlap::NIndexes
