+++ ydb/core/tx/columnshard/engines/storage/indexes/bloom/meta.cpp
@@ -1,4 +1,4 @@
-#include "checker.h"
+#include "bits_storage.h"
 #include "meta.h"
 
 #include <ydb/core/formats/arrow/hash/calcer.h>
