+++ ydb/core/tx/columnshard/engines/storage/indexes/bloom/ya.make
@@ -3,7 +3,7 @@ LIBRARY()
 SRCS(
     GLOBAL constructor.cpp
     GLOBAL meta.cpp
-    GLOBAL checker.cpp
+    bits_storage.cpp
 )
 
 PEERDIR(
