+++ ydb/core/tx/columnshard/engines/storage/indexes/bloom_ngramm/ya.make
@@ -3,7 +3,6 @@ LIBRARY()
 SRCS(
     GLOBAL constructor.cpp
     GLOBAL meta.cpp
-    GLOBAL checker.cpp
     const.cpp
 )
