+++ ydb/core/tx/columnshard/engines/storage/indexes/count_min_sketch/meta.cpp
@@ -1,5 +1,4 @@
 #include "meta.h"
-#include "checker.h"
 #include <ydb/library/formats/arrow/hash/xx_hash.h>
 #include <ydb/core/formats/arrow/hash/calcer.h>
 #include <ydb/core/tx/program/program.h>
