+++ ydb/core/tx/columnshard/engines/storage/indexes/count_min_sketch/ya.make
@@ -3,7 +3,6 @@ LIBRARY()
 SRCS(
     GLOBAL constructor.cpp
     GLOBAL meta.cpp
-    GLOBAL checker.cpp
 )
 
 PEERDIR(
